Add safety around font parameters
Issue : The `getFontParameters` function can return `null`, and the current
implementation is not guaranteed to return an item. This cause issues when there
are label that need to be printed.

Fix : We add the font size in the default font string (as it is necessary to the
regex). We also ensure we have default values in case the regex within `getFontParameters` does not manage to parse anything (for example, it can't parse numbers in font names)